better implementation of Arc.intersect(Line), plus tests
This implementation provides special handling in Arc::intersect()
when `self` is a non-rotated Arc and `other_seg` is a Line.  In this
case it uses the straight-forward closed-form solution to identify the
intersection points.

Rotated Arcs and Arcs intersecting with non-Line objects still use
the pre-existing intersection code, that part is totally untouched by
this commit.